requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
httpx>=0.27.0
python-dotenv>=1.0.0
numpy>=2.1.0
torch>=2.1.0
torchaudio>=2.1.0
soundfile>=0.12.1
nemo_toolkit[asr]>=1.23.0
requests>=2.31.0
pyannote.audio>=3.1.0
desktop-notifier>=6.0.0
sentry-sdk>=1.0.0
langfuse>=3.0.0
